pyproject.toml
# https://flit.pypa.io/en/stable/pyproject_toml.html#pyproject-project-dependencies
[build-system]
requires = ["flit_core >=3.8,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "flow4df"
version = "0.0.1"
authors = [
  { name="Boyan Georgiev", email="todo" },
]
description = "My dscription"
readme = "README.md"
requires-python = ">=3.11,<4"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "flit>=3.8",
    "black",
    "pendulum>=2.1.2",
    "pyspark==3.5.1",
    "delta-spark==3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
]
doc = [
    "sphinx>=7.3.0",
]
emacs = [
    "ipython",
    "pyright==1.1.345",
]

[project.urls]
Documentation = "TODO"
Source = "TODO"

[tool.pytest.ini_options]
markers = [
    "slow: Slow tests",
]
